Optimize size of lookup table for certain CRCs
The needed lookup table entry size is a function of the
*polynomial's* bit width (ignoring the implied top bit), *not*
the CRC's width. If the polynomial's width is smaller than the
CRC's width, we can in certain cases save some space.

We don't apply the optimization to reflected CRCs because,
unlike for unreflected CRCs, that would introduce more work
in the main loop.